fix(cli): fix security enhancement notification for projects with schema dir
